Fix(runner): thread array does not resize even if allocated (#144)
`runner_impl` uses `set` which does not support resize; therefore,
resize logic was added iff the thread array is full.

fixes: #142

+++ inkcpp_test/FallbackFunction.cpp
@@ -22,7 +22,7 @@ SCENARIO("run a story with external function and fallback function", "[external 
 			int  cnt_sqrt = 0;
 			auto fn_sqrt  = [&cnt_sqrt](int x) -> int {
         ++cnt_sqrt;
-        return sqrt(x);
+				return static_cast<int>(sqrt(x));
 			};
 			int  cnt_greeting = 0;
 			auto fn_greeting  = [&cnt_greeting]() -> const char* {
@@ -51,7 +51,7 @@ SCENARIO("run a story with external function and fallback function", "[external 
 			int  cnt_sqrt = 0;
 			auto fn_sqrt  = [&cnt_sqrt](int x) -> int {
         ++cnt_sqrt;
-        return sqrt(x);
+				return static_cast<int>(sqrt(x));
 			};
 
 			thread->bind("sqrt", fn_sqrt);
